+++ src/lang-xpath/test/uk/co/reecedunn/intellij/plugin/xpath/tests/parser/XPathParserTest.kt
@@ -3597,7 +3597,7 @@ class XPathParserTest : ParserTestCase() {
     internal inner class PostfixExpr_Lookup {
         @Test
         @DisplayName("key specifier; NCName")
-        fun keySpecifier_ncname() {
+        fun ncname() {
             val expected = loadResource("tests/parser/xpath-3.1/Lookup_NCName.txt")
             val actual = parseResource("tests/parser/xpath-3.1/Lookup_NCName.xq")
             assertThat(actual.toPsiTreeString(), `is`(expected))
@@ -3605,7 +3605,7 @@ class XPathParserTest : ParserTestCase() {
 
         @Test
         @DisplayName("key specifier; IntegerLiteral")
-        fun keySpecifier_IntegerLiteral() {
+        fun integerLiteral() {
             val expected = loadResource("tests/parser/xpath-3.1/Lookup_IntegerLiteral.txt")
             val actual = parseResource("tests/parser/xpath-3.1/Lookup_IntegerLiteral.xq")
             assertThat(actual.toPsiTreeString(), `is`(expected))
@@ -3613,15 +3613,31 @@ class XPathParserTest : ParserTestCase() {
 
         @Test
         @DisplayName("key specifier; ParenthesizedExpr")
-        fun keySpecifier_ParenthesizedExpr() {
+        fun parenthesizedExpr() {
             val expected = loadResource("tests/parser/xpath-3.1/Lookup_ParenthesizedExpr.txt")
             val actual = parseResource("tests/parser/xpath-3.1/Lookup_ParenthesizedExpr.xq")
             assertThat(actual.toPsiTreeString(), `is`(expected))
         }
 
         @Test
+        @DisplayName("key specifier; ParenthesizedExpr; missing VarName")
+        fun parenthesizedExpr_missingVarName() {
+            val expected = loadResource("tests/parser/xpath-3.1/Lookup_ParenthesizedExpr_MissingVarName.txt")
+            val actual = parseResource("tests/parser/xpath-3.1/Lookup_ParenthesizedExpr_MissingVarName.xq")
+            assertThat(actual.toPsiTreeString(), `is`(expected))
+        }
+
+        @Test
+        @DisplayName("key specifier; ParenthesizedExpr; missing VarName; multiple")
+        fun parenthesizedExpr_missingVarName_multiple() {
+            val expected = loadResource("tests/parser/xpath-3.1/Lookup_ParenthesizedExpr_MissingVarName_Multiple.txt")
+            val actual = parseResource("tests/parser/xpath-3.1/Lookup_ParenthesizedExpr_MissingVarName_Multiple.xq")
+            assertThat(actual.toPsiTreeString(), `is`(expected))
+        }
+
+        @Test
         @DisplayName("key specifier; wildcard")
-        fun keySpecifier_Wildcard() {
+        fun keySpecifier_wildcard() {
             val expected = loadResource("tests/parser/xpath-3.1/Lookup_Wildcard.txt")
             val actual = parseResource("tests/parser/xpath-3.1/Lookup_Wildcard.xq")
             assertThat(actual.toPsiTreeString(), `is`(expected))
@@ -3909,7 +3925,7 @@ class XPathParserTest : ParserTestCase() {
     internal inner class UnaryLookup {
         @Test
         @DisplayName("key specifier; NCName")
-        fun keySpecifier_ncname() {
+        fun ncname() {
             val expected = loadResource("tests/parser/xpath-3.1/UnaryLookup_NCName.txt")
             val actual = parseResource("tests/parser/xpath-3.1/UnaryLookup_NCName.xq")
             assertThat(actual.toPsiTreeString(), `is`(expected))
@@ -3917,7 +3933,7 @@ class XPathParserTest : ParserTestCase() {
 
         @Test
         @DisplayName("key specifier; IntegerLiteral")
-        fun keySpecifier_IntegerLiteral() {
+        fun integerLiteral() {
             val expected = loadResource("tests/parser/xpath-3.1/UnaryLookup_IntegerLiteral.txt")
             val actual = parseResource("tests/parser/xpath-3.1/UnaryLookup_IntegerLiteral.xq")
             assertThat(actual.toPsiTreeString(), `is`(expected))
@@ -3925,15 +3941,31 @@ class XPathParserTest : ParserTestCase() {
 
         @Test
         @DisplayName("key specifier; ParenthesizedExpr")
-        fun keySpecifier_ParenthesizedExpr() {
+        fun parenthesizedExpr() {
             val expected = loadResource("tests/parser/xpath-3.1/UnaryLookup_ParenthesizedExpr.txt")
             val actual = parseResource("tests/parser/xpath-3.1/UnaryLookup_ParenthesizedExpr.xq")
             assertThat(actual.toPsiTreeString(), `is`(expected))
         }
 
         @Test
+        @DisplayName("key specifier; ParenthesizedExpr; missing VarName")
+        fun parenthesizedExpr_missingVarName() {
+            val expected = loadResource("tests/parser/xpath-3.1/UnaryLookup_ParenthesizedExpr_MissingVarName.txt")
+            val actual = parseResource("tests/parser/xpath-3.1/UnaryLookup_ParenthesizedExpr_MissingVarName.xq")
+            assertThat(actual.toPsiTreeString(), `is`(expected))
+        }
+
+        @Test
+        @DisplayName("key specifier; ParenthesizedExpr; missing VarName; multiple")
+        fun parenthesizedExpr_missingVarName_multiple() {
+            val expected = loadResource("tests/parser/xpath-3.1/UnaryLookup_ParenthesizedExpr_MissingVarName_Multiple.txt")
+            val actual = parseResource("tests/parser/xpath-3.1/UnaryLookup_ParenthesizedExpr_MissingVarName_Multiple.xq")
+            assertThat(actual.toPsiTreeString(), `is`(expected))
+        }
+
+        @Test
         @DisplayName("key specifier; wildcard")
-        fun keySpecifier_Wildcard() {
+        fun wildcard() {
             val expected = loadResource("tests/parser/xpath-3.1/UnaryLookup_Wildcard.txt")
             val actual = parseResource("tests/parser/xpath-3.1/UnaryLookup_Wildcard.xq")
             assertThat(actual.toPsiTreeString(), `is`(expected))