+++ CHANGELOG.md
@@ -25,6 +25,7 @@ XPath and XQuery:
 
 1. Fix parsing `FTContainsExpr` in the RHS of a `ComparisonExpr` in XPath expressions.
 2. Improve the error message when a `return` is used without a `ForExpr` or `LetExpr` in XPath expressions.
+3. Don't generate an error when parsing `?$` and `?($)`, where the `VarName` is missing.
 
 ## 2021